add landmarks to 3dmm improvements to fitter and light related functions
